Major change: The WsLoop can now receive messages as well as send them to the connected user.
Minor change: Re-factored the Ws checking, handshake building etc. to
another module (mochiweb_ws_utils). This is because I plan to change
the web socket server into a gen_server at another date.

Problems: For a weird reason the first message received from the user is
interpreted as a byte array, but every request after that is shown as a
string. This is probably a socket setting problem that I haven't figured
out how to get past.